+++ lanebatchupdate.py
@@ -17,6 +17,7 @@ import os
 import re
 
 from .lane_fix_controller import LaneFixController
+from .excel_preview_controller import ExcelPreviewController
 from .reconstruct_controller import ReconstructController
 
 
@@ -34,6 +35,7 @@ class LaneBatchUpdateTool:
         self.field_names = {}
         self.reconstruct = ReconstructController(iface, self.plugin_dir, self.log)
         self.lane_fix = LaneFixController(iface, self.plugin_dir, self.log)
+        self.excel_preview = ExcelPreviewController(iface, self.plugin_dir, self.log)
 
     def initGui(self):
         buttons = (
@@ -50,6 +52,7 @@ class LaneBatchUpdateTool:
             self.actions.append(action)
         self.reconstruct.initGui(self.actions)
         self.lane_fix.initGui(self.actions)
+        self.excel_preview.initGui(self.actions)
 
     def unload(self):
         for action in self.actions:
@@ -58,6 +61,7 @@ class LaneBatchUpdateTool:
         self.actions = []
         self.reconstruct.unload()
         self.lane_fix.unload()
+        self.excel_preview.unload()
 
     @staticmethod
     def is_empty(value):